+++ primitives/domain/src/lib.rs
@@ -229,7 +229,7 @@ impl TryFrom<Vec<u8>> for MainchainPublicKey {
 pub const MAINCHAIN_ADDRESS_HASH_LEN: usize = 28;
 
 /// Some hash of MainchainAddress, 28 bytes. Presumably blake2b_224.
-/// Way to get it: cardano-cli address key-hash --payment-verification-key-file <path to vkey>
+/// Way to get it: cardano-cli address key-hash --payment-verification-key-file FILE
 #[derive(
 	Clone, Copy, Decode, Default, Eq, Encode, Hash, MaxEncodedLen, PartialEq, ToDatum, TypeInfo,
 )]
@@ -524,7 +524,7 @@ impl CandidateRegistrations {
 /// -- keyi - 33 bytes compressed ECDSA public key of a committee member
 /// -- @
 /// newtype ATMSPlainAggregatePubKey = ATMSPlainAggregatePubKey ByteString
-/// https://github.com/input-output-hk/partner-chains-smart-contracts/blob/5b19d25a95c3ab49ae0e4c6ce0ec3376f13b3766/docs/Specification.md#L554-L561
+/// <https://github.com/input-output-hk/partner-chains-smart-contracts/blob/5b19d25a95c3ab49ae0e4c6ce0ec3376f13b3766/docs/Specification.md#L554-L561>
 #[derive(Clone, Debug, Decode, Encode, Eq, PartialEq, ToDatum, TypeInfo)]
 pub struct ATMSPlainAggregatePubKey(pub [u8; 32]);